FORGE-1307 Fixed dependency resolution errors against secure repos
This ports the FORGE-1134 fix to Forge 2. It ensures that
dependency resolution includes any all repositories in all active
profiles, any mirrors to such repositories and uses proxies to
access any repositories if configured in the Maven settings.xml
file.

This is not a direct port of the bugfix. It uses the newer Eclipse
Aether APIs.