FORGE-1321 Resolve Authentication before Aether request execution.
Mirror detection of repositories is performed to ensure that
mirrors in the form of RemoteRepository objects are returned
instead of the real repositories.

Additionally, Authentication info is enriched after the final list
of mirrors and repositories are obtained. This ensures that
CollectRequests and VersionRangeRequests are supplied with the
right set of mirrors and repositories.